@@ -2,24 +2,27 @@ package com.vladimircvetanov.smartfinance.accounts;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.vladimircvetanov.smartfinance.IFragment;
 import com.vladimircvetanov.smartfinance.R;
 import com.vladimircvetanov.smartfinance.db.DBAdapter;
 import com.vladimircvetanov.smartfinance.favourites.AddCategoryDialogFragment;
 import com.vladimircvetanov.smartfinance.favourites.IconsAdapter;
 import com.vladimircvetanov.smartfinance.model.Manager;
 import com.vladimircvetanov.smartfinance.model.RowDisplayable;
+import com.vladimircvetanov.smartfinance.transactionRelated.ItemType;
 
 import java.util.ArrayList;
 
-public class AccountsFragment extends Fragment {
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
+public class AccountsFragment extends Fragment implements IFragment {
 
     private RecyclerView accountsList;
     private RecyclerView moreAccountIconsList;
@@ -42,7 +45,7 @@ public class AccountsFragment extends Fragment {
         final ArrayList<RowDisplayable> accounts = new ArrayList<>();
         accounts.addAll(adapter.getCachedAccounts().values());
 
-        accountsAdapter = new AccountsAdapter(accounts, getActivity(), getFragmentManager());
+        accountsAdapter = new AccountsAdapter(accounts, getActivity(), getParentFragmentManager());
         accountsList = (RecyclerView) view.findViewById(R.id.accounts_list);
         accountsList.setAdapter(accountsAdapter);
         accountsList.setLayoutManager(new LinearLayoutManager(getActivity()));
@@ -65,11 +68,16 @@ public class AccountsFragment extends Fragment {
                         arguments.putString("ROW_DISPLAYABLE_TYPE", "ACCOUNT");
 
                         dialog.setArguments(arguments);
-                        dialog.show(getFragmentManager(), String.valueOf(R.string.add_category_dialog));
-
+                        dialog.show(getParentFragmentManager(), String.valueOf(R.string.add_category_dialog));
                     }
                 })
         );
         return view;
+    }
+
+    @Override
+    public void updateData(RowDisplayable item, ItemType type, Boolean delete) {
+        accountsAdapter.addAccount(item);
+        accountsAdapter.notifyDataSetChanged();
     }
 }
