@@ -1,8 +1,6 @@
 package com.vladimircvetanov.smartfinance.favourites;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.DialogFragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -10,10 +8,17 @@ import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.vladimircvetanov.smartfinance.MainActivity;
 import com.vladimircvetanov.smartfinance.R;
 import com.vladimircvetanov.smartfinance.db.DBAdapter;
 import com.vladimircvetanov.smartfinance.message.Message;
 import com.vladimircvetanov.smartfinance.model.CategoryExpense;
+import com.vladimircvetanov.smartfinance.transactionRelated.ItemType;
+
+import java.util.concurrent.TimeUnit;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.DialogFragment;
 
 /**
  * Created by vladimircvetanov on 01.05.17.
@@ -26,7 +31,11 @@ public class AddToFavDeleteDialogFragment extends DialogFragment {
     private Button addToFav;
     private Button cancel;
 
+    private IUpdateData updateData;
+
     DBAdapter adapter;
+
+    public AddToFavDeleteDialogFragment() {}
 
     @Nullable
     @Override
@@ -37,6 +46,8 @@ public class AddToFavDeleteDialogFragment extends DialogFragment {
         cancel = (Button) view.findViewById(R.id.cancel_category_button);
         deleteCategory = (Button) view.findViewById(R.id.delete_category_button);
         addToFav = (Button) view.findViewById(R.id.add_to_fav_button);
+
+        updateData = (MainActivity) getActivity();
 
 
         adapter = DBAdapter.getInstance(getActivity());
@@ -61,6 +72,7 @@ public class AddToFavDeleteDialogFragment extends DialogFragment {
                     if(adapter.getCachedExpenseCategories().size() > 1) {
                         adapter.moveToFav(cat);
                         Toast.makeText(getActivity(), "Category added to favorites!", Toast.LENGTH_SHORT).show();
+                        updateData.sendData(cat, ItemType.FAVOURITECATEGORY, false);
                     } else{
                         Message.message(getActivity(),"You can`t be without categories!");
                     }
@@ -76,7 +88,7 @@ public class AddToFavDeleteDialogFragment extends DialogFragment {
             public void onClick(View v) {
                if(adapter.getCachedExpenseCategories().size() > 1) {
                    adapter.deleteExpenseCategory(cat);
-
+                   updateData.sendData(cat, ItemType.CATEGORY, true);
                    Message.message(getActivity(),"Category deleted!");
                }else{
                    Message.message(getActivity(),"You can`t be without categories!");
@@ -87,7 +99,6 @@ public class AddToFavDeleteDialogFragment extends DialogFragment {
         });
         return view;
     }
-
 }
 
 
