@@ -1,8 +1,6 @@
 package com.vladimircvetanov.smartfinance.favourites;
 
 import android.app.Activity;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -12,6 +10,9 @@ import android.widget.ImageView;
 import com.vladimircvetanov.smartfinance.R;
 
 import java.util.ArrayList;
+
+import androidx.core.content.ContextCompat;
+import androidx.recyclerview.widget.RecyclerView;
 
 public class IconsAdapter extends RecyclerView.Adapter<IconsAdapter.IconViewHolder>{
 
