@@ -1,8 +1,6 @@
 package com.vladimircvetanov.smartfinance.reports;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.DialogFragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -19,6 +17,9 @@ import com.vladimircvetanov.smartfinance.model.Account;
 
 import java.util.ArrayList;
 import java.util.HashSet;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.DialogFragment;
 
 public class AccountSelectionDialog extends DialogFragment {
 
